final exercises

+++ part 4/bloglist/controllers/blogs.js
@@ -1,21 +1,9 @@
 const express = require("express");
-const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
 
 const Blog = require("../models/blog");
-const User = require("../models/user");
-const config = require("../utils/config");
 
 const blogsRouter = express.Router();
-
-const getTokenFrom = request => {
-  const authorization = request.get("authorization");
-
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "");
-  }
-
-  return null;
-};
 
 blogsRouter.get("/", async (_req, res) => {
   const blogs = await Blog.find({}).populate("user", {
@@ -26,19 +14,8 @@ blogsRouter.get("/", async (_req, res) => {
   res.json(blogs);
 });
 
-blogsRouter.post("/", async (req, res) => {
-  try {
-    const token = getTokenFrom(req);
-
-    const decodedToken = jwt.verify(token, config.SECRET);
-
-    if (!decodedToken.id) {
-      return res.status(401).json({
-        error: "token invalid",
-      });
-    }
-
-    const user = await User.findById(decodedToken.id);
+blogsRouter.post("/", middleware.userExtractor, async (req, res) => {
+    const user = req.user;
 
     const blog = new Blog({
       ...req.body,
@@ -51,13 +28,6 @@ blogsRouter.post("/", async (req, res) => {
     await user.save();
 
     res.status(201).json(savedBlog);
-  } catch (error) {
-    console.error(error);
-
-    res.status(401).json({
-      error: "token missing or invalid",
-    });
-  }
 });
 
 blogsRouter.put("/:id", async (req, res) => {
@@ -73,7 +43,29 @@ blogsRouter.put("/:id", async (req, res) => {
   res.json(updatedBlog);
 });
 
-blogsRouter.delete("/:id", async (req, res) => {
+blogsRouter.delete("/:id", middleware.userExtractor, async (req, res) => {
+  const user = req.user;
+
+  const blog = await Blog.findById(req.params.id);
+
+  if (!blog) {
+    return res.status(404).json({
+      error: "blog not found",
+    });
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return res.status(401).json({
+      error: "only the creator can delete this blog",
+    });
+  }
+
+  user.blogs = user.blogs.filter(
+    blogId => blogId.toString() !== blog._id.toString()
+  );
+
+  await user.save();
+
   await Blog.findByIdAndDelete(req.params.id);
 
   res.status(204).end();
